chore(lint): enable react-refresh/only-export-components
Move Paths constant from Routes.tsx to Constants.ts and extract
useRouteParams hook to its own file so Routes.tsx only exports the
AppRoutes component. Add inline disable for the intentional
provider+hook co-location in useFeatureFlags.tsx.

Ref: https://github.com/securesign/rhtas-console-ui/issues/307
