+++ client/src/app/Routes.tsx
@@ -1,25 +1,18 @@
 import { Suspense, lazy } from "react";
 import { ErrorBoundary } from "react-error-boundary";
-import { Navigate, useParams, useRoutes } from "react-router-dom";
+import { Navigate, useRoutes } from "react-router-dom";
 
 import { Bullseye, Spinner } from "@patternfly/react-core";
 import { ErrorFallback } from "./components/ErrorFallback";
 import NotFound from "./pages/NotFound";
 import { useFeatureFlags } from "./hooks/useFeatureFlags";
+import { Paths } from "./Constants";
 
 const Artifacts = lazy(() => import("./pages/Artifacts"));
 const TrustRoot = lazy(() => import("./pages/TrustRoot"));
 const RekorSearch = lazy(() => import("./pages/Rekor/RekorSearch"));
 const RekorEntry = lazy(() => import("./pages/Rekor/RekorEntry"));
 const Monitoring = lazy(() => import("./pages/Monitoring"));
-
-export const Paths = {
-  artifacts: "/artifacts",
-  rekorSearch: "/rekor-search",
-  rekorEntry: `/rekor-search/:logIndex`,
-  trustRoot: "/trust-root",
-  monitoringAlerting: "/monitoring",
-} as const;
 
 export const AppRoutes = () => {
   const { features } = useFeatureFlags();
@@ -47,13 +40,4 @@ export const AppRoutes = () => {
       </ErrorBoundary>
     </Suspense>
   );
-};
-
-export const useRouteParams = (pathParam: "logIndex") => {
-  const params = useParams();
-  const value = params[pathParam];
-  if (value === undefined) {
-    throw new Error(`ASSERTION FAILURE: required path parameter not set: ${pathParam}`);
-  }
-  return value;
 };
